+++ html_form_builder_snippets/doc/changelog.rst
@@ -1,3 +1,7 @@
+v.1.7.1
+=======
+* Fail safe required validation
+
 v.1.7
 =====
 * Datetime generation